separated into 5 components
