+++ src/components/Summary.js
@@ -2,25 +2,28 @@ import React, { Component } from "react";
 
 
 class Summary extends Component {
-
-    summary = Object.keys(this.state.selected).map((feature, idx) => {
-        const featureHash = feature + '-' + idx;
-        const selectedOption = this.state.selected[feature];
-        let USCurrencyFormat = new Intl.NumberFormat('en-US', {
-          style: 'currency',
-          currency: 'USD'
-        });
-        return (
-          <div className="summary__option" key={featureHash}>
-            <div className="summary__option__label">{feature} </div>
-            <div className="summary__option__value">{selectedOption.name}</div>
-            <div className="summary__option__cost">
-              {USCurrencyFormat.format(selectedOption.cost)}
+  render () {  
+    const summary = Object.keys(this.props.selected).map((feature, idx) => {
+          const featureHash = feature + '-' + idx;
+          const selectedOption = this.props.selected[feature];
+          let USCurrencyFormat = new Intl.NumberFormat('en-US', {
+            style: 'currency',
+            currency: 'USD'
+          });
+          return (
+            <div className="summary__option" key={featureHash}>
+              <div className="summary__option__label">{feature} </div>
+              <div className="summary__option__value">{selectedOption.name}</div>
+              <div className="summary__option__cost">
+                {USCurrencyFormat.format(selectedOption.cost)}
+              </div>
             </div>
-          </div>
-        );
-      });
-
+          );
+        })
+    return (
+    <div>{summary}</div>
+    )
+  }
 }
 
 export default Summary;
